# webapi/docs/swagger.yaml
definitions:
  dto.ArmorDetailResponse:
    properties:
      armorType:
        type: string
      armorValue:
        type: integer
      durability:
        type: integer
      slot:
        type: string
    type: object
  dto.ItemDetailResponse:
    properties:
      armor:
        $ref: '#/definitions/dto.ArmorDetailResponse'
      id:
        type: integer
      item_type:
        type: string
      max_stack:
        type: integer
      name:
        type: string
      price:
        type: number
      quest_id:
        type: integer
      quest_name:
        type: string
      rarity:
        type: string
      required_level:
        type: integer
      weapon:
        $ref: '#/definitions/dto.WeaponDetailResponse'
    type: object
  dto.ItemListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Item'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.ListRequest:
    properties:
      filters:
        additionalProperties: true
        type: object
      page:
        type: integer
      page_size:
        type: integer
      search:
        type: string
      sort_by:
        type: string
      sort_order:
        description: '"asc" или "desc"'
        type: string
    type: object
  dto.NpcDetailResponse:
    properties:
      faction:
        type: string
      health:
        type: integer
      id:
        type: integer
      level:
        type: integer
      location:
        type: string
      mana:
        type: integer
      name:
        type: string
      reaction:
        type: string
      respawn_time:
        type: integer
      tameable:
        type: boolean
    type: object
  dto.NpcListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Npc'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.QuestDetailResponse:
    properties:
      difficulty:
        type: integer
      id:
        type: integer
      is_repeatable:
        type: boolean
      name:
        type: string
      quest_level:
        type: integer
      required_character_level:
        type: integer
      reward_money:
        type: integer
    type: object
  dto.QuestListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Quest'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  dto.Response:
    properties:
      content: {}
      message:
        type: string
      status:
        type: integer
    type: object
  dto.UserCreateRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  dto.UserLoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  dto.UserLoginResponse:
    properties:
      acess_token:
        type: string
      refresh_token:
        type: string
    type: object
  dto.WeaponDetailResponse:
    properties:
      damage:
        type: string
      durability:
        type: integer
      slot:
        type: string
      speed:
        format: float32
        type: number
      weaponType:
        type: string
    type: object
  enums.ItemTypeEnum:
    enum:
    - Оружие
    - Броня
    - Расходуемые
    type: string
    x-enum-varnames:
    - WeaponType
    - ArmorType
    - OtherType
  enums.RarityEnum:
    enum:
    - Обычный
    - Необычный
    - Редкий
    - Эпический
    - Легендарный
    type: string
    x-enum-varnames:
    - CommonRarity
    - UncommonRarity
    - RareEnumRarity
    - EpicRarity
    - LegendaryRarity
  models.Item:
    properties:
      id:
        type: integer
      itemType:
        $ref: '#/definitions/enums.ItemTypeEnum'
      maxStack:
        type: integer
      name:
        type: string
      price:
        type: number
      questId:
        type: integer
      questName:
        type: string
      rarity:
        $ref: '#/definitions/enums.RarityEnum'
      requiredLevel:
        type: integer
    type: object
  models.Npc:
    properties:
      faction:
        type: string
      health:
        type: integer
      id:
        type: integer
      level:
        type: integer
      location:
        type: string
      mana:
        type: integer
      name:
        type: string
      reaction:
        type: string
      respawnTime:
        type: integer
      tameable:
        type: boolean
    type: object
  models.Quest:
    properties:
      difficulty:
        type: integer
      id:
        type: integer
      name:
        type: string
      questLevel:
        type: integer
      repeatable:
        type: boolean
      requiredCharacterLevel:
        type: integer
      rewardMoney:
        type: integer
    type: object
  models.User:
    properties:
      hash:
        type: string
      id:
        type: integer
      login:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/item/{id}:
    get:
      description: Возвращает предмет по уникальному ID. Может возвращать базовый
        предмет (Item), оружие (Weapon) или броню (Armor) в зависимости от типа предмета
      parameters:
      - description: ID предмета
        in: path
        name: id
        required: true
        type: integer
      - description: 'Access token в формате: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Базовый предмет
          schema:
            $ref: '#/definitions/dto.ItemDetailResponse'
        "400":
          description: Неверный формат ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Предмет не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить предмет по ID
      tags:
      - Предметы
  /api/item/list:
    post:
      consumes:
      - application/json
      description: 'Возвращает список предметов с поддержкой пагинации, сортировки,
        фильтрации и поиска. Поддерживает различные типы предметов: базовые предметы
        (Item), оружие (Weapon) и броню (Armor)'
      parameters:
      - description: Параметры запроса для фильтрации, сортировки и пагинации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ListRequest'
      - description: 'Access token в формате: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком предметов
          schema:
            $ref: '#/definitions/dto.ItemListResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Предметы не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список предметов с фильтрацией и пагинацией
      tags:
      - Предметы
  /api/npc/{id}:
    get:
      description: Возвращает НИП по уникальному ID
      parameters:
      - description: ID нип
        in: path
        name: id
        required: true
        type: integer
      - description: 'Access token в формате: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: НИП
          schema:
            $ref: '#/definitions/dto.NpcDetailResponse'
        "400":
          description: Неверный формат ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: НИП не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить НИП по ID
      tags:
      - НИП
  /api/npc/list:
    post:
      consumes:
      - application/json
      description: Возвращает список НИП с поддержкой пагинации, сортировки, фильтрации
        и поиска.
      parameters:
      - description: Параметры запроса для фильтрации, сортировки и пагинации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ListRequest'
      - description: 'Access token в формате: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком НИП
          schema:
            $ref: '#/definitions/dto.NpcListResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: НИП не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список НИП с фильтрацией и пагинацией
      tags:
      - НИП
  /api/quest/{id}:
    get:
      description: Возвращает квест по уникальному ID
      parameters:
      - description: ID квест
        in: path
        name: id
        required: true
        type: integer
      - description: 'Access token в формате: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: квест
          schema:
            $ref: '#/definitions/dto.QuestDetailResponse'
        "400":
          description: Неверный формат ID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: квест не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить квест по ID
      tags:
      - Квесты
  /api/quest/list:
    post:
      consumes:
      - application/json
      description: Возвращает список квестов с поддержкой пагинации, сортировки, фильтрации
        и поиска.
      parameters:
      - description: Параметры запроса для фильтрации, сортировки и пагинации
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.ListRequest'
      - description: 'Access token в формате: Bearer {token}'
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ со списком квестов
          schema:
            $ref: '#/definitions/dto.QuestListResponse'
        "400":
          description: Неверный формат запроса
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Квест не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить список квестов с фильтрацией и пагинацией
      tags:
      - Квесты
  /api/user:
    post:
      consumes:
      - application/json
      description: Создаёт нового пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Пользователь успешно создан
          schema:
            $ref: '#/definitions/dto.Response'
        "400":
          description: Неверный формат данных или пользователь уже существует
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Создать пользователя
      tags:
      - users
  /api/user/{id}:
    get:
      consumes:
      - application/json
      description: Возвращает пользователя по уникальному ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Получить пользователя по ID
      tags:
      - users
  /api/users:
    post:
      consumes:
      - application/json
      description: Авторизирует пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/dto.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Пользователь успешно авторизован
          schema:
            $ref: '#/definitions/dto.UserLoginResponse'
        "400":
          description: Неверный формат данных
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Авторизация пользователя
      tags:
      - users
swagger: "2.0"